Added solution

+++ 02. Python Fundamentals/01. Basic Syntax, Conditional Statements and Loops - Lab/1.2. Basic Syntax, Conditional Statements and Loops - Exercise/03. Chat Codes.py
@@ -1,7 +1,7 @@
 n = int(input())
 
 for i in range(n):
-    number = float(input())
+    number = int(input())
     if number == 88:
         print('Hello')
     elif number == 86:
